+++ examples/bevy-basic-sculpt/src/systems/update.rs
@@ -1,7 +1,7 @@
 use bevy::picking::backend::ray::RayMap;
 use bevy::prelude::*;
-use freestyle_sculpt::SculptParams;
 use freestyle_sculpt::ray::Ray;
+use freestyle_sculpt::SculptParams;
 use mesh_graph::MeshGraph;
 
 use crate::resources::*;