+++ src/selectors/mod.rs
@@ -2,6 +2,8 @@ mod sphere_with_falloff;
 mod surface_sphere_with_falloff;
 mod traits;
 
+use hashbrown::HashSet;
+use mesh_graph::{FaceId, MeshGraph, VertexId};
 pub use sphere_with_falloff::*;
 pub use surface_sphere_with_falloff::*;
 pub use traits::*;
@@ -33,4 +35,19 @@ fn get_sphere_with_falloff_weight_callback(
             0.0
         }
     })
+}
+
+fn faces_incident_to_vertices(
+    vertices: impl IntoIterator<Item = VertexId>,
+    mesh_graph: &MeshGraph,
+) -> HashSet<FaceId> {
+    let mut faces = HashSet::new();
+
+    for vertex_id in vertices {
+        for face in mesh_graph.vertices[vertex_id].faces(mesh_graph) {
+            faces.insert(face);
+        }
+    }
+
+    faces
 }