+++ src/selectors/surface_sphere_with_falloff.rs
@@ -3,7 +3,10 @@ use hashbrown::HashSet;
 
 use mesh_graph::{Face, MeshGraph, Selection};
 
-use super::{get_sphere_with_falloff_weight_callback, MeshSelector, WeightedSelection};
+use super::{
+    MeshSelector, WeightedSelection, faces_incident_to_vertices,
+    get_sphere_with_falloff_weight_callback,
+};
 
 /// Generates a selection on the surface of a mesh that is within a sphere with a falloff and that
 /// is limited to be connected to the input face.
@@ -68,7 +71,7 @@ impl MeshSelector for SurfaceSphereWithFalloff {
 
         WeightedSelection {
             selection: Selection {
-                vertices,
+                faces: faces_incident_to_vertices(vertices, mesh_graph),
                 ..Default::default()
             },
             get_weight: get_sphere_with_falloff_weight_callback(